+++ kaparoo/filesystem/hierarchy/__init__.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+__all__ = (
+    "Directory",
+    "Entry",
+    "Expandable",
+    "File",
+    "Literal",
+    "Template",
+)
+
+from kaparoo.filesystem.hierarchy.nodes import Directory, Entry, File
+from kaparoo.filesystem.hierarchy.patterns import Expandable, Literal, Template